Improve hyperparameter optimization
Move run_hyper out of the package.
